# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SABABUKA Bersinar API
  version: 1.0.0-draft
  summary: Kontrak API MVP untuk tata kelola data Kabupaten Kapuas
  description: |
    API tunggal untuk ruang Superadmin, BAPPERIDA, Kominfo, OPD, Pimpinan,
    dashboard eksekutif, dan Asisten AI. Semua operasi dibatasi oleh permission,
    lingkup organisasi, status workflow, dan feature flag.
  contact:
    name: Pengelola SABABUKA Bersinar
  license:
    name: Penggunaan Internal Pemerintah Kabupaten Kapuas
    identifier: LicenseRef-SABABUKA-Internal
servers:
  - url: http://127.0.0.1:3001/api/v1
    description: Local development (aktif saat ini)
  - url: https://dev.sababuka.com/api/v1
    description: Development (belum diaktifkan)
  - url: https://sababuka.com/api/v1
    description: Production (setelah disetujui)
...
security:
  - sessionCookie: []
paths:
...
  /submissions:
    get:
      tags: [Submissions]
      operationId: listSubmissions
      summary: Daftar pengiriman data
      description: Membutuhkan `submission.view`; hasil mengikuti scope.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: organization_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/SubmissionStatus"}
        - name: period_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
      responses:
        "200":
          description: Daftar submission
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionPage"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Submissions]
      operationId: createSubmission
      summary: Membuat batch pengiriman draft
      description: Membutuhkan `submission.create`.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SubmissionCreateInput"}
      responses:
        "201":
          description: Batch draft dibuat
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetail"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /submissions/{submission_id}:
    get:
      tags: [Submissions]
      operationId: getSubmission
      summary: Rincian form capaian beserta indikator aktif yang ditugaskan
      description: Membutuhkan `submission.view`; akses mengikuti scope organisasi.
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      responses:
        "200":
          description: Rincian form capaian
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetail"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /submissions/{submission_id}/observations/{indicator_version_id}:
    put:
      tags: [Submissions]
      operationId: saveSubmissionObservation
      summary: Menyimpan realisasi satu indikator pada draft
      description: Membutuhkan `submission.update`; hanya berlaku untuk indikator aktif yang ditugaskan kepada OPD.
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: indicator_version_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ObservationInput"}
      responses:
        "200":
          description: Realisasi tersimpan
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionObservation"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/InvalidTransition"}
...
  /submissions/{submission_id}/evidence:
    get:
      tags: [Submissions]
      operationId: listSubmissionEvidence
      summary: Daftar bukti dukung privat sesuai scope
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
      responses:
        "200": {description: Daftar metadata bukti dukung}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Submissions]
      operationId: uploadSubmissionEvidence
      summary: Unggah PDF, JPG, PNG, atau XLSX maksimum 10 MB
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: indicator_version_id
          in: query
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        "201": {description: Bukti tersimpan dengan checksum SHA-256}
        "413": {$ref: "#/components/responses/UploadTooLarge"}
        "415": {$ref: "#/components/responses/UnsupportedFile"}
  /submissions/{submission_id}/evidence/{evidence_id}:
    delete:
      tags: [Submissions]
      operationId: deleteSubmissionEvidence
      summary: Soft-delete bukti dukung pada form yang masih dapat diubah
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: evidence_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": {description: Bukti dihapus}
        "409": {$ref: "#/components/responses/InvalidTransition"}
  /submissions/{submission_id}/evidence/{evidence_id}/download:
    get:
      tags: [Submissions]
      operationId: downloadSubmissionEvidence
      summary: Unduh bukti dukung privat sesuai scope
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: evidence_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/Uuid"}
      responses:
        "200": {description: Isi berkas bukti dukung}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /submissions/{submission_id}/actions/{action}:
    post:
      tags: [Submissions]
      operationId: transitionSubmission
      summary: Menjalankan transisi workflow pengiriman
      description: |
        Form input manual mendukung `submit`, `start-review`, `return`, dan `approve`.
        Jalur upload berkas akan memakai `validate` dan `confirm` pada tahap berikutnya.
      parameters:
        - $ref: "#/components/parameters/SubmissionId"
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [submit, start-review, return, approve]
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TransitionInput"}
      responses:
        "200":
          description: Transisi sinkron berhasil
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetail"}
        "202":
          description: Validasi berjalan asinkron
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetail"}
        "400": {$ref: "#/components/responses/InvalidTransition"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/OpenValidationErrors"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sababuka_session
      description: Sesi server-side; cookie wajib HttpOnly, Secure, dan SameSite=Lax.
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 20}
      description: Token CSRF terikat ke sesi untuk seluruh operasi mutasi.
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    SubmissionId:
      name: submission_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Uuid"}
...
  responses:
...
    Forbidden:
      description: Permission atau scope tidak mencukupi
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    NotFound:
      description: Objek tidak ditemukan atau tidak boleh diungkapkan
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    ValidationError:
      description: Permintaan tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    InvalidTransition:
      description: Transisi workflow tidak valid
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    OpenValidationErrors:
      description: Error validasi masih terbuka
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
    UploadTooLarge:
      description: Berkas terlalu besar
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
    UnsupportedFile:
      description: Format berkas tidak didukung
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiError"}
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    PageMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1}
        total_items: {type: integer, minimum: 0}
        total_pages: {type: integer, minimum: 0}
    PageEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {}}
        meta: {$ref: "#/components/schemas/PageMeta"}
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {$ref: "#/components/schemas/Uuid"}
            details:
              type: array
              items:
                type: object
                additionalProperties: true
...
    SubmissionStatus:
      type: string
      enum: [draft, validating, ready, submitted, opd_confirmed, under_review, returned, approved, published, cancelled]
    SubmissionCreateInput:
      type: object
      additionalProperties: false
      required: [organization_id, period_id]
      properties:
        organization_id: {$ref: "#/components/schemas/Uuid"}
        period_id: {$ref: "#/components/schemas/Uuid"}
    SubmissionDetail:
      type: object
      required: [id, dataset_version_id, organization_id, reporting_period_id, submission_method, status, row_count, created_at]
      properties:
        id: {$ref: "#/components/schemas/Uuid"}
        dataset_version_id: {$ref: "#/components/schemas/Uuid"}
        organization_id: {$ref: "#/components/schemas/Uuid"}
        organization_code: {type: string}
        organization_name: {type: string}
        reporting_period_id: {$ref: "#/components/schemas/Uuid"}
        period_code: {type: string}
        period_label: {type: string}
        submission_method: {type: string}
        source_filename: {type: [string, "null"]}
        status: {$ref: "#/components/schemas/SubmissionStatus"}
        row_count: {type: integer, minimum: 0}
        validation_summary:
          type: object
          properties:
            errors: {type: integer, minimum: 0}
            warnings: {type: integer, minimum: 0}
            information: {type: integer, minimum: 0}
        created_at: {$ref: "#/components/schemas/Timestamp"}
        updated_at: {$ref: "#/components/schemas/Timestamp"}
        observations:
          type: array
          items: {$ref: "#/components/schemas/SubmissionObservation"}
    ObservationInput:
      type: object
      additionalProperties: false
      required: [value]
      properties:
        value:
          oneOf:
            - type: number
            - type: string
              minLength: 1
              maxLength: 4000
        notes: {type: [string, "null"], maxLength: 4000}
    SubmissionObservation:
      type: object
      required: [indicator_version_id, indicator_code, indicator_name, data_type]
      properties:
        observation_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        indicator_version_id: {$ref: "#/components/schemas/Uuid"}
        indicator_code: {type: string}
        indicator_name: {type: string}
        data_type: {type: string}
        unit_name: {type: string}
        unit_symbol: {type: [string, "null"]}
        target_numeric_value: {type: [number, "null"]}
        target_text_value: {type: [string, "null"]}
        numeric_value: {type: [number, "null"]}
        text_value: {type: [string, "null"]}
        notes: {type: [string, "null"]}
        quality_status: {type: [string, "null"]}
    SubmissionPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: {$ref: "#/components/schemas/SubmissionDetail"}
    TransitionInput:
      type: object
      additionalProperties: false
      properties:
        notes: {type: [string, "null"], maxLength: 4000}
...